chkstat: move assorted types into dedicated header
